requirements.txt
streamlit==0.69.2
pandas==1.1.3
matplotlib==3.3.2
numpy==1.19.2
python_dateutil==2.8.1
